+++ config/routes.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 Emailbutler::Engine.routes.draw do
-  post '/webhooks', to: 'webhooks#create_deprecated'
-  post '/webhooks/:provider', to: 'webhooks#create'
+  post '/webhooks/:provider', to: 'webhooks#create' if Emailbutler.configuration.providers.any?
 
   resources :ui, only: %i[index show]
   namespace :ui do