+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.2] - 2024-06-16
+### Modified
+- removed deprecated webhooks action
+- webhooks route optional if no provider present
+
+## [0.8.1] - 2024-06-04
 ### Added
 - resend provider integration
 - mailjet provider integration
+
+### Modified
+- make providers configuration optional
 
 ## [0.8.0] - 2024-05-11
 ### Added